requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
praw==7.6.0
prawcore==2.3.0
python-dotenv==0.20.0
python-telegram-bot==13.12
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.0
six==1.16.0
tornado==6.1
tzdata==2022.1
tzlocal==4.2
update-checker==0.18.0
urllib3==1.26.9
websocket-client==1.3.3
Werkzeug==2.1.2
